backend/requirements-dev.txt
# =================================================================
# Development Dependencies for SoutiAI Transcription Engine
# =================================================================
# These are additional packages needed for development, testing, and debugging
# Install with: pip install -r requirements-dev.txt

# =================================================================
# Testing Frameworks
# =================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0
pytest-html>=4.1.0
pytest-benchmark>=4.0.0

# =================================================================
# Code Quality & Linting
# =================================================================
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.4.0

# =================================================================
# Documentation
# =================================================================
mkdocs>=1.5.0
mkdocs-material>=9.1.0
mkdocstrings>=0.22.0

# =================================================================
# Development Tools
# =================================================================
jupyter>=1.0.0
ipykernel>=6.23.0
notebook>=6.5.0

# =================================================================
# Debugging & Profiling
# =================================================================
debugpy>=1.6.0
memory-profiler>=0.61.0
line-profiler>=4.1.0
py-spy>=0.3.0

# =================================================================
# API Testing
# =================================================================
httpx>=0.24.0
requests>=2.31.0
responses>=0.23.0

# =================================================================
# Mocking & Fixtures
# =================================================================
faker>=19.0.0
factory-boy>=3.2.0

# =================================================================
# Database Testing
# =================================================================
pytest-postgresql>=5.0.0
pytest-redis>=3.0.0

# =================================================================
# Load Testing
# =================================================================
locust>=2.15.0

# =================================================================
# Security Testing
# =================================================================
bandit>=1.7.0
safety>=2.3.0

# =================================================================
# Environment Management
# =================================================================
python-dotenv>=1.0.0

# =================================================================
# Logging & Monitoring (Development)
# =================================================================
rich>=13.5.0
loguru>=0.7.0

# =================================================================
# Type Checking
# =================================================================
# types-all>=1.0.0  # Deprecated, use specific types packages
types-requests>=2.31.0
# types-redis>=4.5.0  # May have compatibility issues

# =================================================================
# Development Servers
# =================================================================
uvicorn[standard]>=0.23.0
watchfiles>=0.19.0

# =================================================================
# Database Tools
# =================================================================
alembic>=1.11.0
sqlalchemy-utils>=0.41.0

# =================================================================
# AI/ML Development Tools
# =================================================================
torchinfo>=1.8.0
transformers[torch]>=4.31.0
datasets>=2.14.0
accelerate>=0.21.0

# =================================================================
# Web Development
# =================================================================
jinja2>=3.1.0
aiofiles>=23.1.0

# =================================================================
# Utilities
# =================================================================
click>=8.1.0
tqdm>=4.65.0
psutil>=5.9.0
python-multipart>=0.0.6

# =================================================================
# Cloud Development (Optional)
# =================================================================
# boto3>=1.28.0          # AWS SDK
# google-cloud-storage>=2.10.0  # GCP Storage
# azure-storage-blob>=12.17.0   # Azure Storage

# =================================================================
# Monitoring Development
# =================================================================
prometheus-client>=0.17.0
opentelemetry-distro>=0.40b0
opentelemetry-instrumentation-fastapi>=0.40b0

# =================================================================
# Testing Data
# =================================================================
# These can be large, install as needed
# librosa>=0.10.0        # Audio processing for tests
# moviepy>=1.0.0         # Video processing for tests
# pillow>=10.0.0         # Image processing for tests